event/agents/operasyon/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
pydantic==2.9.2
jinja2==3.1.4
python-multipart==0.0.12
openpyxl==3.1.5
anthropic==0.40.0
python-dotenv==1.0.1
aiofiles==24.1.0
